:traffic_light: Replacing Grunt with Gulp
In an effort to modernize the project, I'm taking out Grunt (the Gruntfile and it's npm dependencies) and replacing it with Gulp. I'm still using Jasmine (for now), but instead of using grunt-contrib-jasmine I'm using Karma to run Phantom and run the jasmine tests in there.

I'm also changing the License from WTFPL to MIT, since this is something that Gilt uses and MIT won't creep as many people out.

Also adding a CONTRIBUTING.md to document the testing setup.